+++ NEWS.md
@@ -1,10 +1,13 @@
+# mini007 0.5.0
 
+- Adding the `per-station retry` features to the `Workflow` class
+- Adding the `fallback handlers` to the `Workflow` class
+
+
+# mini007 0.4.0
 #### Implementing the `Workflow` class which allows one to have full control over a `workflow` of agents
 #### Adding new methods:
 - `share_context_with()`
-
-
-#### Creating a quarto website
 
 # mini007 0.3.0
 #### Adding new methods to work with `tools`: